+++ tests/fixes/node-fixes/class-dewidow-fix-test.php
@@ -26,6 +26,7 @@ namespace PHP_Typography\Tests\Fixes\Node_Fixes;
 
 use \PHP_Typography\Fixes\Node_Fixes\Dewidow_Fix;
 use \PHP_Typography\Settings;
+use \PHP_Typography\U;
 
 /**
  * Dewidow_Fix unit test.
@@ -61,22 +62,31 @@ class Dewidow_Fix_Test extends Node_Fix_Testcase {
 	 */
 	public function provide_dewidow_data() {
 		return [
-			[ 'bla foo b', 'bla foo&nbsp;b', 3, 2 ],
-			[ 'bla foo&thinsp;b', 'bla foo&thinsp;b', 3, 2 ], // don't replace thin space...
-			[ 'bla foo&#8202;b', 'bla foo&#8202;b', 3, 2 ],   // ... or hair space.
-			[ 'bla foo bar', 'bla foo bar', 2, 2 ],
-			[ 'bla foo bär...', 'bla foo&nbsp;b&auml;r...', 3, 3 ],
-			[ 'bla foo bär&hellip;!', 'bla foo&nbsp;b&auml;r&hellip;!', 3, 3 ],
-			[ 'bla foo bär&shy;!', 'bla foo&nbsp;b&auml;r!', 3, 3 ],
-			[ 'bla foo b&shy;är!', 'bla foo&nbsp;b&auml;r!', 3, 3 ],
-			[ 'bla foo b&#8203;är!', 'bla foo&nbsp;b&auml;r!', 3, 3 ],
-			[ 'bla foo&nbsp;bär...', 'bla foo&nbsp;b&auml;r...', 3, 3 ],
-			[ 'bla föö&#8203;bar s', 'bla f&ouml;&ouml;&#8203;bar&nbsp;s', 3, 2 ],
-			[ 'bla foo&#8203;bar s', 'bla foo&#8203;bar s', 2, 2 ],
-			[ 'bla foo&shy;bar', 'bla foo&shy;bar', 3, 3 ], // &shy; not matched.
-			[ 'bla foo&shy;bar bar', 'bla foo&shy;bar&nbsp;bar', 3, 3 ], // &shy; not matched, but syllable after is.
-			[ 'bla foo&#8203;bar bar', 'bla foo&#8203;bar&nbsp;bar', 3, 3 ],
-			[ 'bla foo&nbsp;bar bar', 'bla foo&nbsp;bar bar', 3, 3 ], // widow not replaced because the &nbsp; would pull too many letters from previous.
+			[ 'bla foo b', 'bla foo&nbsp;b', 3, 2, 1 ],
+			[ 'bla foo&thinsp;b', 'bla foo&thinsp;b', 3, 2, 1 ], // don't replace thin space...
+			[ 'bla foo&#8202;b', 'bla foo&#8202;b', 3, 2, 1 ],   // ... or hair space.
+			[ 'bla foo bar', 'bla foo bar', 2, 2, 1 ],
+			[ 'bla foo bär...', 'bla foo&nbsp;b&auml;r...', 3, 3, 1 ],
+			[ 'bla foo bär&hellip;!', 'bla foo&nbsp;b&auml;r&hellip;!', 3, 3, 1 ],
+			[ 'bla foo bär&shy;!', 'bla foo&nbsp;b&auml;r!', 3, 3, 1 ],
+			[ 'bla foo b&shy;är!', 'bla foo&nbsp;b&auml;r!', 3, 3, 1 ],
+			[ 'bla foo b&#8203;är!', 'bla foo&nbsp;b&auml;r!', 3, 3, 1 ],
+			[ 'bla foo&nbsp;bär...', 'bla foo&nbsp;b&auml;r...', 3, 3, 1 ],
+			[ 'bla föö&#8203;bar s', 'bla f&ouml;&ouml;&#8203;bar&nbsp;s', 3, 2, 1 ],
+			[ 'bla foo&#8203;bar s', 'bla foo&#8203;bar s', 2, 2, 1 ],
+			[ 'bla foo&shy;bar', 'bla foo&shy;bar', 3, 3, 1 ], // &shy; not matched.
+			[ 'bla foo&shy;bar bar', 'bla foo&shy;bar&nbsp;bar', 3, 3, 1 ], // &shy; not matched, but syllable after is.
+			[ 'bla foo&#8203;bar bar', 'bla foo&#8203;bar&nbsp;bar', 3, 3, 1 ],
+			[ 'bla foo&nbsp;bar bar', 'bla foo&nbsp;bar bar', 3, 3, 1 ], // widow not replaced because the &nbsp; would pull too many letters from previous.
+			[ 'bla foo bar bar', 'bla foo bar&nbsp;bar', 3, 3, 1 ],
+			[ 'bla foo bar bar', 'bla foo bar&nbsp;bar', 3, 3, 2 ],
+			[ 'bla foo bar bar', 'bla foo bar&nbsp;bar', 3, 3, 3 ],
+			[ 'bla foo bar bar', 'bla foo bar&nbsp;bar', 3, 7, 1 ],
+			[ 'bla foo bar bar', 'bla foo&nbsp;bar&nbsp;bar', 3, 7, 2 ],
+			[ 'bla foo bar bar', 'bla foo&nbsp;bar&nbsp;bar', 3, 7, 3 ],
+			[ 'bla bla foo bar bar', 'bla bla foo bar&nbsp;bar', 3, 11, 1 ],
+			[ 'bla bla foo bar bar', 'bla bla foo&nbsp;bar&nbsp;bar', 3, 11, 2 ],
+			[ 'bla bla foo bar bar', 'bla bla&nbsp;foo&nbsp;bar&nbsp;bar', 3, 11, 3 ],
 		];
 	}
 
@@ -84,18 +94,24 @@ class Dewidow_Fix_Test extends Node_Fix_Testcase {
 	 * Test apply.
 	 *
 	 * @covers ::apply
+	 * @covers ::dewidow
+	 *
+	 * @uses ::make_space_nonbreaking
+	 * @uses ::strip_breaking_characters
 	 *
 	 * @dataProvider provide_dewidow_data
 	 *
-	 * @param string $html       HTML input.
-	 * @param string $result     Expected result.
-	 * @param int    $max_pull   Maximum number of pulled characters.
-	 * @param int    $max_length Maximum word length for dewidowing.
+	 * @param string $html        HTML input.
+	 * @param string $result      Expected result.
+	 * @param int    $max_pull    Maximum number of pulled characters.
+	 * @param int    $max_length  Maximum word length for dewidowing.
+	 * @param int    $word_number Maximum number of words in widow.
 	 */
-	public function test_apply( $html, $result, $max_pull, $max_length ) {
+	public function test_apply( $html, $result, $max_pull, $max_length, $word_number ) {
 		$this->s->set_dewidow( true );
 		$this->s->set_max_dewidow_pull( $max_pull );
 		$this->s->set_max_dewidow_length( $max_length );
+		$this->s->set_dewidow_word_number( $word_number );
 
 		$this->assertFixResultSame( $html, $result );
 	}
@@ -104,19 +120,47 @@ class Dewidow_Fix_Test extends Node_Fix_Testcase {
 	 * Test dewidow.
 	 *
 	 * @covers ::apply
+	 * @covers ::dewidow
+	 *
+	 * @uses ::make_space_nonbreaking
+	 * @uses ::strip_breaking_characters
 	 *
 	 * @dataProvider provide_dewidow_data
 	 *
-	 * @param string $html       HTML input.
-	 * @param string $result     Expected result.
-	 * @param int    $max_pull   Maximum number of pulled characters.
-	 * @param int    $max_length Maximum word length for dewidowing.
+	 * @param string $html        HTML input.
+	 * @param string $result      Expected result.
+	 * @param int    $max_pull    Maximum number of pulled characters.
+	 * @param int    $max_length  Maximum word length for dewidowing.
+	 * @param int    $word_number Maximum number of words in widow.
 	 */
-	public function test_apply_off( $html, $result, $max_pull, $max_length ) {
+	public function test_apply_off( $html, $result, $max_pull, $max_length, $word_number ) {
 		$this->s->set_dewidow( false );
 		$this->s->set_max_dewidow_pull( $max_pull );
 		$this->s->set_max_dewidow_length( $max_length );
+		$this->s->set_dewidow_word_number( $word_number );
 
 		$this->assertFixResultSame( $html, $html );
+	}
+
+	/**
+	 * Test strip_breaking_characters.
+	 *
+	 * @covers ::strip_breaking_characters
+	 */
+	public function test_strip_breaking_characters() {
+		$result = $this->invokeStaticMethod( Dewidow_Fix::class, 'strip_breaking_characters', [ 'foo' . U::SOFT_HYPHEN . 'bar' . U::ZERO_WIDTH_SPACE . 'baz' . U::ZERO_WIDTH_SPACE ] );
+
+		$this->assertSame( 'foobarbaz', $result );
+	}
+
+	/**
+	 * Test make_space_nonbreaking.
+	 *
+	 * @covers ::make_space_nonbreaking
+	 */
+	public function test_make_space_nonbreaking() {
+		$result = $this->invokeStaticMethod( Dewidow_Fix::class, 'make_space_nonbreaking', [ 'foo' . U::SOFT_HYPHEN . 'bar ' . U::ZERO_WIDTH_SPACE . ' baz' . U::ZERO_WIDTH_SPACE, 'u' ] );
+
+		$this->assertSame( 'foo' . U::SOFT_HYPHEN . 'bar' . U::NO_BREAK_SPACE . U::ZERO_WIDTH_SPACE . U::NO_BREAK_SPACE . 'baz' . U::ZERO_WIDTH_SPACE, $result );
 	}
 }